==> src/markdown/bookdown/01_inleiding.Rmd <==
# Inleiding{#inleiding}

```{r setup, message = FALSE, warning = FALSE}
# Opties
library(knitr)
opts_chunk$set(
  echo = TRUE,
  message = FALSE,
  warning = FALSE
)

# Packages
library(tidyverse)
library(sf)
library(mapview)
library(rgbif)

# Source
slak_in_du_dir <- "C:/Users/lange/Documents/onderzoek/Slak-in-Du/analyses/slak-in-du"
source(file.path(slak_in_du_dir, "src", "R", "functions.R"))

# References
knitr::write_bib(c("tidyverse", "sf", "mapview", "rgbif"), 
                 file = "packages.bib")
```

Dit technisch rapport is geschreven ter ondersteuning van het Slak-in-Du rapport na 11 jaar veldwerk in de kustduinen en -polders. Het doel is een volledig beeld van alle mogelijke informatie uit de data weer te geven. Relevante informatie en/of figuren kunnen dan geselecteerd worden die in het eigenlijke rapport moeten worden gebruikt.\
Bovendien wordt de methodiek zo goed mogelijk uitgeschreven zodat mensen met verschillende achtergronden kunnen opzoeken en begrijpen hoe de dataverwerking en -analyses zijn verricht, zonder zelf op de hoogte te moeten zijn van de technische details of gebruikte programmeercode.

Het technisch rapport bestaat uit vier delen:

1.  Afbakening van het Slak-in-Du gebied aan de kust. Het studiegebied kan dan gebruikt worden om de data te selecteren en UTM-hokken te selecteren die kunnen gebruikt worden om verspreidingskaarten te maken. Hoofdstuk \@ref(studiegebied).
2.  Selectie van de data en data preparatie voor verdere analyses. Hoofdstuk \@ref(dataselectie).
3.  Maken van "atlas-stijl" verspreidingskaarten m.b.v. UTM-hokken op verschillende schalen. Hoofdstuk \@ref(verspreidingskaarten).
4.  Exploratie van de data om samenvattende cijfers en tabellen te genereren. Hoofdstuk \@ref(exploratie).

Alle dataverwerking en analyses zijn verricht met R v4.3.3 in RStudio v2023.12.1.402 [@RStudio2024;@R2024] tenzij anders vermeld. Voor dataverwerking en visualisatie werd gebruik gemaakt van de **tidyverse** package v2.0.0 [@R-tidyverse; @tidyverse2019]. De **sf** package v1.0-15 [@R-sf; @sf2018] werd gebruikt voor de verwerking van de kaartlagen en **mapview** v2.11.2 [@R-mapview] voor het creëren van interactieve kaarten. Ten slotte werd de **rgbif** package v3.7.9 [@R-rgbif; @rgbif2017] gebruikt voor het ophalen van de weteschappelijke namen via de GBIF taxonomic backbone [@GBIFtax2023, via GBIF.org op `r Sys.Date()`].
