==> src/markdown/bookdown/03_dataselectie_kust.Rmd <==
# Dataselectie{#dataselectie}
## Doel

Selectie van de observaties die gebruikt zullen in het eindrapport van Slak-in-Du aan de kust. Voorbereiden van finale dataset voor data exploratie en analyses.


## Data inlezen

We lezen de volgende kaarten in: 

- Studiegebied aan de kust

Deze bevat twee ecoregio's. We voegen deze samen zodat we een enkele kaart hebben van het volledige studiegebied. 

```{r dataselectie-data}
# Locatie van data en output mappen
data_path <- file.path(slak_in_du_dir, "data", "raw")
out_path <- file.path(slak_in_du_dir, "data", "processed")

# Lees kaartlagen in
studiegebied_kust <- st_read(paste(out_path, "studiegebied_kust.shp", 
                                   sep = "/"))

# Voeg ecoregio's samen tot één studiegebied
studiegebied_kust_merged <- studiegebied_kust %>%
  st_buffer(0.5) %>%
  st_union()

# Referentie coordinatensysteem: UTM zone 31N
proj_crs <- 32631
```

Visualisatie van de kaartlagen:

```{r dataselectie-kaarten}
# Maak thema voor plotten van kaartjes
mytheme <- theme(panel.grid.major = element_line(color = "grey", 
                                                 linetype = "dashed", 
                                                 size = 0.3),
                 panel.background = element_rect(fill = "aliceblue"),
                 axis.title = element_blank())

# Plot kaarten
studiegebied_kust_merged %>%
  ggplot() +
    geom_sf(fill = "white") +
    mytheme
```

We lezen de waarnemingendata in van Slak-in-Du. Laatste download 21-01-2023.

```{r dataselectie-waarnemingen}
# Laad data
ruwe_data <- read_csv(paste(data_path, "observations-slak-in-du.csv",
                               sep = "/"))

# Zet om naar sf object
ruwe_data_sf <- ruwe_data %>% 
  mutate(latitude = lat, longitude = lng) %>%
  st_as_sf(coords = c("longitude", "latitude"), crs = 4326)

# Zet crs om naar project crs
ruwe_data_sf <- ruwe_data_sf %>%
  st_transform(crs = proj_crs)
```

## Dataselectie

We selecteren daarna de waarnemingen binnen het studiegebied aan de kust.

```{r dataselectie-intersect-studiegebied}
# Selectie waarnemingen binnen studiegebied
waarnemingen_kust <- ruwe_data_sf %>%
  st_intersection(studiegebied_kust_merged)
```

### Randgevallen

Zijn er punten die vlak buiten het gebied vallen, maar die wel van belang zijn?

```{r dataselectie-randgevallen}
# Selecteer correcte observaties binnen gebied
correct_obs <- waarnemingen_kust
id_correct_obs <- correct_obs$id

# Selecteer observaties net buiten gebied
net_erbuiten <- ruwe_data_sf %>% 
  st_crop(xmin = 460000, xmax = 530000,
          ymin = 5640000, ymax = 5700000) %>%
  filter(!(id %in% id_correct_obs))

# Visualiseer
mapview(correct_obs, col.regions = "red") +
  mapview(net_erbuiten, col.regions = "blue") +
  mapview(studiegebied_kust_merged, col.regions = "white")
```

De punten ten zuiden van het studiegebied bevinden zich terecht niet in de selectie voor het huidige rapport. Het zijn waarnemingen van excursies naar Sint-Michiels en Ryckevelde/Schobbejakshoogte.

De punten ten noorden van het studiegebied moeten ook niet opgenomen worden. Het gaat om een aangespoelde schelp op het strand van Zeebrugge een een waarneming van een mariene soort (Amerikaanse strandschelp - *Mulinia lateralis*) aan de Baai van Heist.


### Selectie van soorten

We verwijderen alle mariene soorten uit de selectie. Het gaat om:  
Mossel, Japanse oester, Strandgaper, Macoma balthica, Witte dunschaal, Zaagje, Filipijnse tapijtschelp, Grofgeribde fuikhoren, Penhoren, Amerikaanse boormossel, Paalworm, Melkwitte arkschalp, Klein drijfhorentje, Ruwe alikruik-complex, Witte wenteltrap, Platte oester, Witte boormossel, Kokkel, Noorse hartschelp, Halfgeknotte strandschelp, Venusschelp, Gewone alikruik, Grote strandschelp, Angulus tenuis, Roze schotelhoren, Tweetandschelpje, Gekielde cirkelslak, Noordse rotsboorder, Bonte mantel, Gevlekt koffieboontje, Gewone priktolhoren, Purperslak, Stompe ribhoren, Vliezig drijfhorentje, Tolhoren onbekend, Gevlochten fuikhoren, Scheve hartschelp, Scheef bultschelpje, Stompe buishoren, Muizenkeuteltje, Klein traliehorentje, Paardenzadel, Haustator solanderi, Glanzende tepelhoren, Oubliehoren, Gewone marmerschelp, Wulk, Grote tepelhoren, Hongaarse muts, Amerikaanse strandschelp, Corbula gibba, Platte slijkgaper, Tere dunschaal, Driehoekige parelmoerneut.

```{r dataselectie-marien}
mariene_mollusken <- c("Mossel", "Japanse oester", "Strandgaper", 
  "Macoma balthica", "Witte dunschaal", "Zaagje", "Filipijnse tapijtschelp", 
  "Grofgeribde fuikhoren", "Penhoren", "Amerikaanse boormossel", "Paalworm", 
  "Melkwitte arkschelp", "Klein drijfhorentje", "Ruwe alikruik-complex", 
  "Witte wenteltrap", "Platte oester", "Witte boormossel", "Kokkel", 
  "Noorse hartschelp", "Halfgeknotte strandschelp", "Venusschelp", 
  "Gewone alikruik", "Grote strandschelp", "Angulus tenuis", 
  "Roze schotelhoren", "Tweetandschelpje", "Gekielde cirkelslak", 
  "Noordse rotsboorder", "Bonte mantel", "Gevlekt koffieboontje", 
  "Gewone priktolhoren", "Purperslak", "Stompe ribhoren", 
  "Vliezig drijfhorentje", "Tolhoren onbekend", "Gevlochten fuikhoren", 
  "Scheve hartschelp", "Scheef bultschelpje", "Stompe buishoren", 
  "Muizenkeuteltje", "Klein traliehorentje", "Paardenzadel", 
  "Haustator solanderi", "Glanzende tepelhoren", "Oubliehoren", 
  "Gewone marmerschelp", "Wulk", "Grote tepelhoren", "Hongaarse muts", 
  "Amerikaanse strandschelp", "Corbula gibba", "Platte slijkgaper", 
  "Tere dunschaal", "Driehoekige parelmoerneut")

land_zoetwater_kust <- waarnemingen_kust %>% 
  filter(!(`species name` %in% mariene_mollusken))
```


### Selectie relevante waarnemingen


```{r dataselectie-te-verwijderen}
te_verwijderen <- c()
```

Zijn er waarnemingen van *aangespoelde schelpen*?

- in ```activity```

```{r dataselectie-aangespoeld-activity}
land_zoetwater_kust %>%
  st_drop_geometry() %>%
  count(activity) %>%
  arrange(n) %>%
  kable()
```

Er is een aangespoelde schelp van de Grote glansslak en enkele van het Gewoon en Meertandig muizenoortje:

```{r dataselectie-aangespoeld-activity-verwijderen}
land_zoetwater_kust %>%
  st_drop_geometry() %>%
  filter(activity == "aangespoeld") %>% 
  select(date, `species name`, number, `life stage`, activity, location, 
         link) %>%
  kable()

# voeg id toe aan vector om te verwijderen
te_verwijderen <- c(te_verwijderen, 
  land_zoetwater_kust %>%
    st_drop_geometry() %>%
    filter(activity == "aangespoeld") %>% 
    pull(id))
```

- in ```notes```

```{r dataselectie-aangespoeld-notes}
any(grepl("aangespoeld", land_zoetwater_kust$notes, ignore.case = TRUE))
```

Woord 'aangespoeld' niet aanwezig in notes.  
  
  
Zijn er waarnemingen van *fossiele schelpen*?

- in ```method```

```{r dataselectie-fossiel-method}
land_zoetwater_kust %>%
  st_drop_geometry() %>%
  count(method) %>%
  arrange(n) %>%
  kable()
```

1 fossiel opgezwollen brakwaterhorentje en 4 fossiele wadslakjes:

```{r dataselectie-fossiel-method-verwijderen}
land_zoetwater_kust %>%
  st_drop_geometry() %>%
  filter(method == "fossiel") %>% 
  select(date, `species name`, number, `life stage`, activity, location, notes,
         link) %>%
  kable()

# voeg id toe aan vector om te verwijderen
te_verwijderen <- c(te_verwijderen, 
  land_zoetwater_kust %>%
    st_drop_geometry() %>%
    filter(method == "fossiel") %>% 
    pull(id))
```

- in ```notes```

$n$ verdere waarnemingen bevatten in notities het woord Holoceen, Pleistoceen, fossiel of 'middeleeuw'. De volgende tabel toont aantal per soort:

```{r dataselectie-fossiel-notes}
land_zoetwater_kust %>%
  st_drop_geometry() %>%
  filter(method != "fossiel",
         grepl("fossiel|holoce(en|ne)|pleistoce(en|ne)|middeleeuw", notes, 
               ignore.case = TRUE)) %>%
  count(`species name`) %>%
  kable()

# voeg id toe aan vector om te verwijderen
te_verwijderen <- c(te_verwijderen, 
  land_zoetwater_kust %>%
    st_drop_geometry() %>%
  filter(method != "fossiel",
         grepl("fossiel|holoce(en|ne)|pleistoce(en|ne)|middeleeuw", notes, 
               ignore.case = TRUE)) %>% 
    pull(id))
```

We verwijderen de waarnemingen van de aangespoelde en fossiele schelpen. Het gaat om `r length(te_verwijderen)` waarnemingen.

```{r dataselectie-verwijder-selectie}
land_zoetwater_kust_selectie <- land_zoetwater_kust %>%
  filter(!(id %in% te_verwijderen))
```


## Data voorbereiden
### Controleer accuraatheid waarnemingen

Er is altijd onzekerheid op de exacte locatie van een waarneming. We zien een groot aantal waarnemingen met een accuraatheid groter dan 100 m (arbitrair maximum bepaald na communicatie via mail):

```{r dataselectie-precisie-waarnemingen}
land_zoetwater_kust_selectie %>%
  st_drop_geometry() %>%
  count(accuracy) %>%
  arrange(accuracy) %>%
  kable("html") %>%
  kableExtra::kable_styling(full_width = FALSE)
```

De waarnemingen van 26-05-2018 werden door Hans De Blauwe ingegeven en zetten we voor de zekerheid op 100 m nauwkeurigheid. De andere waarnemingen > 100 m werden door Franky gecontroleerd en zijn correct. Deze zetten we op 25 m. Alle waarnemingen met nauwkeurigheid < 25 m zetten we ook op 25 m (arbitrair minimum).

```{r dataselectie-precisie-verander}
land_zoetwater_kust_selectie <- land_zoetwater_kust_selectie %>%
  mutate(accuracy = case_when(
    accuracy < 25 ~ 25,
    accuracy > 100 & date == "2018-05-26" ~ 100,
    accuracy > 100 & date != "2018-05-26" ~ 25,
    TRUE ~ accuracy
  ))
```

De verdeling is nu als volgt:

```{r dataselectie-precisie-aangepast}
land_zoetwater_kust_selectie %>%
  st_drop_geometry() %>%
  count(accuracy) %>%
  arrange(accuracy) %>%
  kable("html") %>%
  kableExtra::kable_styling(full_width = FALSE)
```

In de data exploratie kunnen we dit visualiseren.


### Wetenschappelijke soortnamen

Gebruik de "GBIF-naamontleder" om nomenclatuurinformatie op te halen voor de wetenschappelijke namen in de dataset.

```{r dataselectie-parse-gbif, cache=TRUE}
parsed_names <- land_zoetwater_kust_selectie %>%
  distinct(`scientific name`) %>%
  pull() %>%
  parsenames()
```

Toon wetenschappelijke namen met nomenclatuurproblemen, d.w.z. niet van het `type = SCIENTIFIC` of die niet volledig zijn ontleed. Let op: deze zijn niet noodzakelijk onjuist.

```{r dataselectie-show-parsed}
probleem_namen <- parsed_names %>%
  select(scientificname, type, parsed, parsedpartially, rankmarker) %>%
  filter(!(type == "SCIENTIFIC" & 
           parsed == "TRUE" & 
           parsedpartially == "FALSE"))
probleem_namen %>% kable
```

We voegen eerst wetenschappelijke namen en taxonomische info toe door te verbinden met de [GBIF taxonomic backbone](https://www.gbif.org/dataset/d7dddbf4-2cf0-4f39-9b2a-bb099caae36c) [@GBIFtax2023, via GBIF.org op `r Sys.Date()`].

```{r dataselectie-getnames-gbif, cache=TRUE}
# Voeg wetenschappelijke namen toe
land_zoetwater_kust_selectie <- land_zoetwater_kust_selectie %>%
  st_drop_geometry() %>%
  rename(name = `scientific name`) %>%
  name_backbone_checklist() %>%
  select(verbatim_name, verbatim_family, scientificName, species, genus, family, 
         class, phylum, kingdom) %>%
  cbind(land_zoetwater_kust_selectie %>%
          select(-c(`scientific name`, family)))

# Krijg correcte namen met auteurs
auteurs <- land_zoetwater_kust_selectie %>%
  pull(species) %>%
  name_backbone_checklist() %>%
  select(scientificName) %>%
  pull()

# Maak aparte kolom voor auteursnamen
land_zoetwater_kust_selectie <- land_zoetwater_kust_selectie %>%
  mutate(scientificName = auteurs) %>%
  rowwise() %>%
  mutate(author = trimws(gsub(species, "", scientificName)))
```

Nu lossen we manueel nog de specifieke problemen van verzamelgroepen op.

```{r dataselectie-verzamelgroepen}
soortgroepen <- c("Arion rufus / ater", probleem_namen$scientificname)

land_zoetwater_kust_selectie <- land_zoetwater_kust_selectie %>%
  # Geen info over soortnaam of auteur als het een verzamelgroep is
  mutate(scientificName = 
           ifelse(verbatim_name %in% soortgroepen, NA, scientificName),
         species = 
           ifelse(verbatim_name %in% soortgroepen, NA, species),
         author = 
           ifelse(verbatim_name %in% soortgroepen, NA, author)) %>%
  # Pas soortnamen aan voor verzamelgroepen
  mutate(wet_soortnaam = case_when(
    verbatim_name == "Anisus leucostoma/spirorbis" ~ 
      "Anisus leucostoma/spirorbis",
    verbatim_name == "Arion (Carinarion) spec." ~ 
      "Arion sylvaticus/circumscriptus",
    verbatim_name == "Arion (Kobeltia) spec." ~ 
      "Arion hortensis/distinctus",
    verbatim_name == "Arion (Mesarion) spec." ~ 
      "Arion fuscus/subfuscus",
    verbatim_name == "Arion rufus / ater" ~ 
      "Arion rufus/ater",
    verbatim_name == "Arion rufus/vulgaris" ~ 
      "Arion rufus/vulgaris",
    verbatim_name == "Columella spec." ~ 
      "Columella edentula/aspera",
    verbatim_name == "Euconulus spec." ~ 
      "Euconulus spec.",
    verbatim_name == "Euglesa/Odhneripisidium/Pisidium spec." ~
      "Euglesa/Odhneripisidium/Pisidium spec.",
    verbatim_name == "Oxychilus spec." ~ 
      "Oxychilus spec.",
    verbatim_name == "Oxyloma spec." ~ 
      "Oxyloma elegans/sarsii",
    verbatim_name == "Stagnicola spec." ~ 
      "Stagnicola spec.",
    verbatim_name == "Succineidae indet." ~ 
      "Succineidae indet.",
    TRUE ~ scientificName
  )) %>%
  # Vul missende data aan op genus level
  mutate(genus = ifelse(verbatim_name == "Columella spec.", 
                        "Columella", genus))
```

Blijkbaar wordt het Meertandig muizenoortje *Myosotella denticulata* (Montagu, 1803) beschouwd als een synoniem van het Gewoon muizenoortje *Myosotella myosotis* (Draparnaud, 1801). Zie [GBIF](https://www.gbif.org/species/4359191) en [MolluscaBase](https://www.molluscabase.org/aphia.php?p=taxdetails&id=139672). Momenteel zal ik deze wel nog als aparte taxa beschouwen, dus dit moet manueel aangepast worden.

```{r dataselectie-namen-manueel}
land_zoetwater_kust_selectie <- land_zoetwater_kust_selectie %>%
  mutate(scientificName = ifelse(verbatim_name == "Myosotella denticulata",
                                 "Myosotella denticulata (Montagu, 1803)",
                                 scientificName),
         species = ifelse(verbatim_name == "Myosotella denticulata",
                                 "Myosotella denticulata",
                                 species),
         author = ifelse(verbatim_name == "Myosotella denticulata",
                                 "(Montagu, 1803)",
                                 author),
         wet_soortnaam = ifelse(verbatim_name == "Myosotella denticulata",
                                 "Myosotella denticulata (Montagu, 1803)",
                                 wet_soortnaam))
```

Hebben we nog missing data op genus level of hoger?

```{r dataselectie-check-ongedetermineerd}
land_zoetwater_kust_selectie %>% 
  filter_at(vars(genus, family, class, phylum, kingdom), 
            any_vars(is.na(.))) %>%
  select(wet_soortnaam, species, genus, family, class, phylum, kingdom) %>%
  kable()
```

Ok.


### Nederlandse soortnamen

We vergelijken de Nederlandse soortnamen met de @deBruyne2015. We passen de volgende namen aan (links: naam in dataset, rechts: naam in de @deBruyne2015):  
  
Boswegslak/Grauwe wegslak = Bos-wegslak / Grauwe wegslak  
Boswegslak = Bos-wegslak  
Egelwegslak = Egel-wegslak  
Arion rufus / ater = Rode wegslak / Duistere wegslak  
Rode wegslak/Spaanse wegslak = Rode wegslak / Spaanse wegslak  
Wormnaaktslak = Grijze wormnaaktslak  
Alinda biplicata biplicata = Grote clausilia  
Gewone driehoeksmossel = Driehoeksmossel  
Gewone wijngaardslak = Wijngaardslak  
Eénbandige grasslak = Eenbandige grasslak  
Grote karthuizerslak = Grote kartuizerslak  
Kleine karthuizerslak = Kleine kartuizerslak  
Gewone poelslak = Grote poelslak  
Glansslak-soort = Glansslak onbekend  
Posthoornslak = Posthorenslak  
Gewone kristalslak = Grote kristalslak  
Tandeloze korfslak/Ruwe korfslak = Tandloze korfslak / Ruwe korfslak  
Cylindrische korfslak = Cilindrische korfslak

```{r dataselectie-ned-soortnaam}
land_zoetwater_kust_selectie <- land_zoetwater_kust_selectie %>%
  mutate(`species name` = recode(`species name`, 
           "Boswegslak/Grauwe wegslak" = "Bos-wegslak / Grauwe wegslak",
           "Boswegslak" = "Bos-wegslak",
           "Egelwegslak" = "Egel-wegslak",
           "Arion rufus / ater" = "Rode wegslak / Duistere wegslak",
           "Rode wegslak/Spaanse wegslak" = "Rode wegslak / Spaanse wegslak",
           "Wormnaaktslak" = "Grijze wormnaaktslak",
           "Alinda biplicata biplicata" = "Grote clausilia",
           "Gewone driehoeksmossel" = "Driehoeksmossel",
           "Gewone wijngaardslak" = "Wijngaardslak",
           "Eénbandige grasslak" = "Eenbandige grasslak",
           "Grote karthuizerslak" = "Grote kartuizerslak",
           "Kleine karthuizerslak" = "Kleine kartuizerslak",
           "Gewone poelslak" = "Grote poelslak",
           "Glansslak-soort" = "Glansslak onbekend",
           "Posthoornslak" = "Posthorenslak",
           "Gewone kristalslak" = "Grote kristalslak",
           "Tandeloze korfslak/Ruwe korfslak" = 
             "Tandloze korfslak / Ruwe korfslak",
           "Cylindrische korfslak" = "Cilindrische korfslak"))
```


### Extra kolom toestand waarneming

In de kolom `life stage` vinden we de volgende informatie:

```{r dataselectie-life-stage}
land_zoetwater_kust_selectie %>%
  count(`life stage`) %>%
  arrange(`life stage`) %>%
  kable()
```

In de kolom `activity` vinden we de volgende informatie:

```{r dataselectie-activity} 
land_zoetwater_kust_selectie %>%
  count(activity) %>%
  arrange(activity) %>%
  kable()
```

Verder zit er ook informatie over elke waarneming in `notes`.

We willen een kolom `toestand` toevoegen met de 3 categoriën. Deze zijn te vinden in verschillende kolommen:

-   levend
    -   `life stage`: "levend dier" OF
    -   `activity`: "levend dier" OF
    -   `notes`: zoek voor woord "levend"
-   lege schelp
    -   `life stage`: ("doublet" OF "lege schelp of losse klep") OF
    -   `activity`: "dood" OF
    -   `notes`: zoek voor woord "vers" OF "oud" OF "dood" OF "doublet" OF "fragment" OF "leeg" OF "kokerjuffer" OF alleenstaande V (en verkeerde schrijfwijzes)
-   onbekend
    -   rest

```{r dataselectie-get-toestand}
land_zoetwater_kust_selectie <- land_zoetwater_kust_selectie %>%
  mutate(toestand = case_when(
    `life stage` == "levend dier" | activity == "levend dier" | 
      grepl(pattern = "levend", tolower(notes)) 
         ~ "levend dier",
    `life stage` %in% c("doublet", "lege schelp of losse klep") | 
      activity == "dood" | 
      grepl(pattern = paste0("vers|o.?ud|dood|doublet|(frag|graf)ment|leeg|",
                             "kokerjuf|\\sv(\\s|\\/)"), tolower(notes)) 
         ~ "lege schelp",
    TRUE ~ "onbekend")
    )
```

```{r dataselectie-summary-toestand} 
land_zoetwater_kust_selectie %>%
  count(toestand) %>%
  arrange(toestand) %>%
  kable()
```


## Preparatie dataset

We selecteren de kolommen die voor ons van belang zijn en geven ze logische namen.

```{r dataselectie-preparatie}
final_df <- land_zoetwater_kust_selectie %>%
  select(id, datum = date, ned_soortnaam = `species name`, wet_soortnaam, 
         aantal = number, levensstadium = `life stage`, activiteit = activity, 
         toestand, methode = method, telmethode = `counting method`, 
         locatie = location, latitude = lat, longitude = lng, 
         lambert_x = `local x`, lambert_y = `local y`, precisie = accuracy, 
         notities = notes, is_zeker = `is certain`, link, verbatim_name, 
         species, author, genus, family, class, phylum, kingdom)
```


## Data exporteren

We slaan de finale dataset van waaremingen op als csv-bestand.

```{r dataselectie-export-waarnemingen}
write_csv(final_df, file = paste(out_path, "waarnemingen_kust.csv", sep = "/"))
```

We slaan ook een lijst op van alle gevonden soorten.

```{r dataselectie-export-soortenlijst}
soorten_lijst <- final_df %>%
  distinct(ned_soortnaam, wet_soortnaam, species, author, genus, 
           family, class, phylum, kingdom) %>% 
  arrange(kingdom, phylum, class, family, wet_soortnaam)

openxlsx::write.xlsx(soorten_lijst, 
  file = paste(out_path, "soortenlijst.xlsx", sep = "/"))
```

We slaan ten slotte een lijst op van de waarnemingen niet gedetermineerd tot op soortniveau en waarnemingen waarvan we de toestand niet weten (levend of lege schelp.)

```{r dataselectie-export-onzeker}
ongedetermineerd <- final_df %>%
  filter(is.na(species))

openxlsx::write.xlsx(ongedetermineerd, 
  file = paste(out_path, "ongedetermineerd.xlsx", sep = "/"))

onbekend <- final_df %>%
  filter(toestand == "onbekend")

openxlsx::write.xlsx(onbekend, 
  file = paste(out_path, "onbekende_toestand.xlsx", sep = "/"))
```
