==> src/markdown/bookdown/04_atlas_kaarten_kust.Rmd <==
# Verspreidingskaarten UTM-niveau{#verspreidingskaarten}
## Doel

Verspreidingskaarten maken voor alle soorten binnen het Slak-in-Dugebied aan de kust op verschillende schalen (1, 2, 5, 10 km).


## Data inlezen en voorbereiding

We lezen de volgende kaarten in: 

- UTM 1 km hokken in het Slak-in-Dugebied aan de kust
- UTM 2 km hokken in het Slak-in-Dugebied aan de kust
- UTM 5 km hokken in het Slak-in-Dugebied aan de kust
- UTM 10 km hokken in het Slak-in-Dugebied aan de kust
- Slak-in-Dugebied aan de kust

```{r verspreiding-data, results="hide"}
# Locatie van data en output mappen
data_path <- file.path(slak_in_du_dir, "data", "raw")
out_path <- file.path(slak_in_du_dir, "data", "processed")

# Lees kaartlagen in
utm_1km_kust <- st_read(paste(out_path, "utm_1km_kust.shp", sep = "/"))
utm_2km_kust <- st_read(paste(out_path, "utm_2km_kust.shp", sep = "/"))
utm_5km_kust <- st_read(paste(out_path, "utm_5km_kust.shp", sep = "/"))
utm_10km_kust <- st_read(paste(out_path, "utm_10km_kust.shp", sep = "/"))

studiegebied_kust <- st_read(paste(out_path, "studiegebied_kust.shp", 
                                   sep = "/"))
studiegebied_kust_merged <- studiegebied_kust %>%
  st_buffer(0.5) %>%
  st_union()

# Referentie coordinatensysteem: UTM zone 31N
proj_crs <- 32631
```

Visualisatie van de kaartlagen:

```{r verspreiding-kaarten}
# Maak thema voor plotten van kaartjes
mytheme <- theme(panel.grid.major = element_line(color = "grey", 
                                                 linetype = "dashed", 
                                                 size = 0.3),
                 panel.background = element_rect(fill = "aliceblue"),
                 axis.title = element_blank(),
                 axis.ticks = element_blank(),
                 axis.text = element_blank())

# Plot kaarten
studiegebied_kust %>%
  ggplot() +
    geom_sf(aes(fill = ecoregio)) +
    mytheme

utm_1km_kust %>%
  ggplot() +
    geom_sf(data = studiegebied_kust_merged, fill = "white") +
    geom_sf(fill = alpha("white", 0)) +
    coord_sf(datum = proj_crs) +
    labs(title = "UTM 1 km - UTM zone 31N") +
    mytheme

utm_2km_kust %>%
  ggplot() +
    geom_sf(data = studiegebied_kust_merged, fill = "white") +
    geom_sf(fill = alpha("white", 0)) +
    coord_sf(datum = proj_crs) +
    labs(title = "UTM 2 km - UTM zone 31N") +
    mytheme

utm_5km_kust %>%
  ggplot() +
    geom_sf(data = studiegebied_kust_merged, fill = "white") +
    geom_sf(fill = alpha("white", 0)) +
    geom_sf_text(aes(label = TAG), size = 1.5) +
    coord_sf(datum = proj_crs) +
    labs(title = "UTM 5 km - UTM zone 31N") +
    mytheme

utm_10km_kust %>%
  ggplot() +
    geom_sf(data = studiegebied_kust_merged, fill = "white") +
    geom_sf(fill = alpha("white", 0)) +
    geom_sf_text(aes(label = TAG), size = 2.5) +
    coord_sf(datum = proj_crs) +
    labs(title = "UTM 10 km - UTM zone 31N") +
    mytheme
```

We lezen de waarnemingendata in van Slak-in-Du na data selectie en preparatie. We trekken een buffer rond elke waarneming die overeenkomt met hun nauwkeurigheid.

```{r verspreiding-waarnemingen}
# Laad data
slakken_data <- read_csv(paste(out_path, "waarnemingen_kust.csv", sep = "/"))

# Zet om naar sf object
slakken_data_sf <- slakken_data %>% 
  st_as_sf(coords = c("longitude", "latitude"), crs = 4326)

# Zet crs om naar project crs
slakken_data_sf <- slakken_data_sf %>%
  st_transform(crs = proj_crs)

# Voeg buffer toe
slakken_data_sf_buffer <- slakken_data_sf %>%
  st_buffer(slakken_data_sf$precisie)
```


## Methodiek

Om de atlaskaarten te maken, kijken we naar de overlap van de UTM-hokken met de buffers. We willen echter dat buffers die meerdere hokken overlappen slechts het hok kleuren waar ze het grootste opppervlakte hebben. Beschouw bijvoorbeeld de volgende punten:

```{r verspreiding-simulatie}
punten <- data.frame(id = 1:2, 
                     x = c(492000,495500),
                     y = c(5678000, 5672000)) %>%
  st_as_sf(coords = c("x", "y"), crs = proj_crs) %>%
  st_buffer(1000)

utm_5km_kust %>%
  filter(startsWith(TAG, "DS97")) %>%
  ggplot() +
    geom_sf(fill = alpha("white", 0)) +
    geom_sf(data = punten) +
    coord_sf(datum = proj_crs) +
    mytheme
```

We willen niet dat alle hokken met overlap gekleurd worden zoals hier:

```{r verspreiding-plot-simulatie-mis}
# Neem intersectie punten met buffer en grid
intersection <- punten %>%
  st_intersection(utm_5km_kust %>%
    filter(startsWith(TAG, "DS97")))

# Lijst met UTM-hokken waar de soort voorkomt
utm_list <- unique(intersection$TAG)

# Add variable to resolution grid if species is present
utm_5km_kust %>%
  filter(startsWith(TAG, "DS97")) %>%
  mutate(occurrence = ifelse(TAG %in% utm_list, "present", "absent")) %>%
  ggplot() +
    geom_sf(data = punten) +
    geom_sf(aes(fill = occurrence), colour = alpha("white", 0)) +
    scale_fill_manual(values = c(alpha("white", 0), alpha("red", 0.4))) +
    geom_sf(data = utm_5km_kust %>% filter(startsWith(TAG, "DS97")),
            fill = alpha("white", 0)) +
    coord_sf(datum = proj_crs) +
    theme(legend.position = "") +
    mytheme
```

We willen wel de hokken met de grootste overlap kleuren, zoals hier:

```{r verspreiding-plot-simulatie-goed}
# Neem intersectie grid en punten met buffer
intersection <- utm_5km_kust %>%
  filter(startsWith(TAG, "DS97")) %>%
  st_intersection(punten) %>%

# Selecteer per plot het grootste deel
  mutate(oppervlakte = st_area(geometry)) %>%
  group_by(id) %>%
  filter(oppervlakte == max(oppervlakte))


# Lijst met UTM-hokken waar de soort voorkomt
utm_list <- unique(intersection$TAG)

# Add variable to resolution grid if species is present
utm_5km_kust %>%
  filter(startsWith(TAG, "DS97")) %>%
  mutate(occurrence = ifelse(TAG %in% utm_list, "present", "absent")) %>%
  ggplot() +
    geom_sf(data = punten) +
    geom_sf(aes(fill = occurrence), colour = alpha("white", 0)) +
    scale_fill_manual(values = c(alpha("white", 0), alpha("red", 0.4))) +
    geom_sf(data = utm_5km_kust %>% filter(startsWith(TAG, "DS97")),
            fill = alpha("white", 0)) +
    coord_sf(datum = proj_crs) +
    theme(legend.position = "") +
    mytheme
```


## Data exploratie
### Zoekinspanning

Voor we de verspreidingskaarten maken, is het belangrijk om een idee te hebben waar er precies is gezocht en met welke inspanning. Dit zal de interpretatie van de kaarten verbeteren. Zoekinspanning wordt hier gemeten als het aantal dagen per UTM-hok. Hokken waar we niet zijn geweest, zijn niet gekleurd (grijs).

```{r verspreiding-inspanning, cache=TRUE}
resoluties <- c(1, 2, 5, 10)

for (resolutie in resoluties) {
  p <- stats_utm_grid(data = slakken_data_sf_buffer, resolutie = resolutie, 
                      variabele = "datum") %>%
     ggplot() +
       geom_sf(data = studiegebied_kust_merged, fill = "white") +
       geom_sf(aes(fill =  n), alpha = 0.8, colour = alpha("white", 0)) +
       scale_fill_gradient(low = "red", high = "green") + 
       geom_sf(data = utm_10km_kust, fill =  alpha("white", 0), size = 0.7) +
       labs(title = "Aantal bezoeken per UTM hok", fill = "Aantal",
            subtitle = paste0("UTM ", resolutie, " km")) +
       coord_sf(datum = proj_crs) +
       mytheme
  
  if (resolutie %in% c(5, 10)) {
    p <- p + geom_sf_text(aes(label = n))
  }
  
  print(p)
}
```

### Soortenrijkdom

We tellen het aantal soorten per kilometerhok. We beschouwen enkel waarnemingen die tot op soort zijn gedetermineerd. Soortengroepen zijn dus ook niet meegerekend. Hokken waar we niet zijn geweest, zijn niet gekleurd (grijs).

```{r verspreiding-soortenrijkdom, cache=TRUE}
resoluties <- c(1, 2, 5, 10)

data <- slakken_data_sf_buffer %>%
  filter(!is.na(species))

for (resolutie in resoluties) {
  p <- stats_utm_grid(data = data, resolutie = resolutie, 
                      variabele = "species") %>%
     ggplot() +
       geom_sf(data = studiegebied_kust_merged, fill = "white") +
       geom_sf(aes(fill =  n), alpha = 0.8, colour = alpha("white", 0)) +
       scale_fill_gradient(low = "red", high = "green") + 
       geom_sf(data = utm_10km_kust, fill =  alpha("white", 0), size = 0.7) +
       labs(title = "Aantal soorten per UTM hok", fill = "Aantal",
            subtitle = paste0("UTM ", resolutie, " km")) +
       coord_sf(datum = proj_crs) +
       mytheme
  
  if (resolutie %in% c(5, 10)) {
    p <- p + geom_sf_text(aes(label = n))
  }
  
  print(p)
}
```

In hokken waar we minder zijn geweest zullen we uiteraard ook minder soorten zien. We zouden een relatieve maat moeten hebben om weer te geven welke hokken de grootste rijkdom hebben. Het aantal soorten per hok delen door het aantal bezoeken per hok is te extreem omdat er soms maar 1x een hok bezocht is en andere keren veel vaker. De controlerende factor zou allicht gevonden kunnen worden door het aantal soorten op de y-as uit te zetten tegenover het aantal bezoeken en te kijken vanaf hoeveel bezoeken de soortenrijkdom afzwakt. Als we die functie kunnen fitten zouden we een correctie kunnen uitvoeren, maar dit zou hier te ver leiden. Bovendien is ons design hier niet op voorbereid om dergelijke maten te berekenen wat zou kunnen leiden tot verkeerde conclusies.


## Verspreidingskaarten

```{r verspreiding-atlas, cache=TRUE}
soorten <- slakken_data_sf_buffer %>% 
  st_drop_geometry() %>%
  arrange(class, family, wet_soortnaam) %>%
  distinct(ned_soortnaam, wet_soortnaam)
resoluties <- c(1, 2)

for (i in seq_len(nrow(soorten))) {
  soortnaam <- soorten$ned_soortnaam[i]
  wet_naam <- soorten$wet_soortnaam[i]
  for (resolutie in resoluties) {
    print(
    stats_utm_grid(data = slakken_data_sf_buffer, soortnaam = soortnaam, 
                   resolutie = resolutie) %>%
      ggplot() +
        geom_sf(data = studiegebied_kust_merged, fill = "white") +
        geom_sf(aes(fill = occurrence), colour = alpha("white", 0), 
                alpha = 0.8) +
        scale_fill_manual(values = c(alpha("green", 0.7), alpha("red", 0.7)),
                          labels = c("aangetroffen", "niet aangetroffen", 
                                     "niet bezocht")) +
        geom_sf(data = utm_10km_kust, fill =  alpha("white", 0), size = 0.7) +
        labs(title = paste0(soortnaam, " - ", wet_naam), fill = "Legende",
             subtitle = paste0("UTM ", resolutie, " km")) +
        coord_sf(datum = proj_crs) +
        theme(legend.justification = c(1, 0), legend.position = c(1, 0)) +
        mytheme
    )
  }
}
```
